cambios al api

+++ OmegaWeb/src/java/db_ws/DatabaseWS.java
@@ -11,6 +11,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 import javax.ejb.Stateless;
 import javax.jws.WebService;
 import javax.jws.WebMethod;
@@ -50,17 +51,29 @@ public class DatabaseWS {
     }
 
     @WebMethod(operationName = "deleteDatabase")
-    public boolean deleteDatabase(@WebParam(name = "dbName") String dbName,
-            @WebParam(name = "user") String user,
-            @WebParam(name = "pw") String pw) {
-        DatabaseAPI dbApi = new DatabaseAPI();
-        return dbApi.deleteDatabase(dbName, user, pw);
+    public boolean deleteDatabase(@WebParam(name = "user_id") int user_id) {
+        User userObject = getRequestUser(user_id);
+        boolean success = false;
+        if (userObject != null) {
+            String dbName = userObject.getDbName();
+            String user = userObject.getDbUser();
+            String pw = userObject.getDbPw();
+            DatabaseAPI dbApi = new DatabaseAPI();
+            success = dbApi.deleteDatabase(dbName, user, pw);
+            if(success){
+                success = removeDatabaseFromUser(user_id);
+            }
+        } else {
+            System.err.println("userObject es nulo");
+        }
+        return success;
     }
 
     @WebMethod(operationName = "openConnection")
     public boolean openConnection(@WebParam(name = "dbName") String dbName,
             @WebParam(name = "user") String user,
-            @WebParam(name = "pw") String pw) {
+            @WebParam(name = "pw") String pw
+    ) {
         DatabaseAPI dbApi = new DatabaseAPI();
         return dbApi.openConnection(dbName, user, pw);
     }
@@ -68,7 +81,8 @@ public class DatabaseWS {
     @WebMethod(operationName = "closeConnection")
     public boolean closeConnection(@WebParam(name = "dbName") String dbName,
             @WebParam(name = "user") String user,
-            @WebParam(name = "pw") String pw) {
+            @WebParam(name = "pw") String pw
+    ) {
         DatabaseAPI dbApi = new DatabaseAPI();
         return dbApi.closeConnection(dbName, user, pw);
     }
@@ -78,20 +92,23 @@ public class DatabaseWS {
             @WebParam(name = "columnNames") String[] columnNames,
             @WebParam(name = "columnTypes") String[] columnTypes,
             @WebParam(name = "nulls") String[] nulls,
-            @WebParam(name = "pk") String[] pk) {
+            @WebParam(name = "pk") String[] pk
+    ) {
         DatabaseAPI dbApi = new DatabaseAPI();
         return dbApi.createTable(dbName, columnNames, columnTypes, nulls, pk);
     }
 
     @WebMethod(operationName = "deleteTable")
-    public boolean deleteTable(@WebParam(name = "tableName") String dbName) {
+    public boolean deleteTable(@WebParam(name = "tableName") String dbName
+    ) {
         DatabaseAPI dbApi = new DatabaseAPI();
         return dbApi.deleteTable(dbName);
     }
 
     @WebMethod(operationName = "addRow")
     public boolean addRow(@WebParam(name = "tableName") String dbName,
-            @WebParam(name = "values") String[] values) {
+            @WebParam(name = "values") String[] values
+    ) {
         DatabaseAPI dbApi = new DatabaseAPI();
         return dbApi.addRow(dbName, values);
     }
@@ -99,7 +116,8 @@ public class DatabaseWS {
     @WebMethod(operationName = "deleteRow")
     public boolean deleteRow(@WebParam(name = "tableName") String dbName,
             @WebParam(name = "columnNames") String[] columnNames,
-            @WebParam(name = "values") String[] values) {
+            @WebParam(name = "values") String[] values
+    ) {
         DatabaseAPI dbApi = new DatabaseAPI();
         return dbApi.deleteRow(dbName, columnNames, values);
     }
@@ -109,7 +127,8 @@ public class DatabaseWS {
             @WebParam(name = "tableName") String dbName,
             @WebParam(name = "selectColumnNames") String[] selectColumnNames,
             @WebParam(name = "whereColumnNames") String[] whereColumnNames,
-            @WebParam(name = "values") String[] values) {
+            @WebParam(name = "values") String[] values
+    ) {
         DatabaseAPI dbApi = new DatabaseAPI();
         return dbApi.select(dbName, selectColumnNames, whereColumnNames, values);
     }
@@ -147,6 +166,27 @@ public class DatabaseWS {
             connection.setConnection();
             String query = "UPDATE USERS SET DB_NAME=?,DB_USER=?,DB_PASSWORD=? WHERE ID=?";
             List<?> params = Arrays.asList(new Object[]{dbName, user, pw, user_id});
+            if (connection.executeQuery(query, params) > 0) {
+                success = true;
+            }
+        } catch (SQLException sqlEx) {
+            System.err.println(sqlEx.getMessage());
+        } finally {
+            try {
+                connection.closeConnection();
+            } catch (SQLException ex) {
+                System.err.println(ex.getMessage());
+            }
+        }
+        return success;
+    }
+    private boolean removeDatabaseFromUser(int user_id){
+        UsersConnection connection = new UsersConnection();
+        boolean success = false;
+        try {
+            connection.setConnection();
+            String query = "UPDATE USERS SET DB_NAME=?,DB_USER=?,DB_PASSWORD=? WHERE ID=?";
+            List<?> params = Arrays.asList(new Object[]{null, null, null, user_id});
             if (connection.executeQuery(query, params) > 0) {
                 success = true;
             }
